Tuoteryhmän voi nyt poistaa vain jos se on tyhjä

+++ application/tuoteryhma/views.py
@@ -32,8 +32,13 @@ def tuoteryhma_form():
 @login_required()
 def tuoteryhma_poista(tuoteryhma_id):
 
-    Tuoteryhma.query.filter_by(id=tuoteryhma_id).delete()
-    db.session().commit()
+    tr = Tuoteryhma.query.filter_by(id=tuoteryhma_id).first()
+    a = Myytava.query.filter_by(tuoteryhma_id=tuoteryhma_id).count()
+    if a>0:
+        render_template("tuoteryhma/tuoteryhma_list.html", tuoteryhmat = Tuoteryhma.query.all())
+    else:
+        Tuoteryhma.query.filter_by(id=tuoteryhma_id).delete()
+        db.session().commit()
 
     return redirect(url_for("tuoteryhmat_index"))
